fix: reference for some string (#211)
* feat: reorgniza code

* feat: re org code and fix bad case ref for equation
